+++ server/src/fhir/json/definitions/capability_statement.rs
@@ -45,9 +45,6 @@ pub struct CapabilityStatementCow<'a>(
 #[serde(rename = "CapabilityStatement")]
 #[serde(rename_all = "camelCase")]
 struct CapabilityStatementHelper {
-    #[serde(with = "FhirVersionDef")]
-    fhir_version: FhirVersion,
-
     name: String,
 
     title: String,
@@ -61,6 +58,9 @@ struct CapabilityStatementHelper {
     kind: KindDef,
 
     implementation: ImplementationDef,
+
+    #[serde(with = "FhirVersionDef")]
+    fhir_version: FhirVersion,
 
     format: Vec<FormatDef>,
 
@@ -138,10 +138,10 @@ enum KindDef {
 
 #[derive(Debug, Serialize)]
 enum FormatDef {
-    #[serde(rename = "application/fhir+xml")]
+    #[serde(rename = "xml")]
     XML,
 
-    #[serde(rename = "application/fhir+json")]
+    #[serde(rename = "json")]
     JSON,
 }